+++ src/Html.php
@@ -1408,6 +1408,7 @@ final class Html
      * Generates a {@see Html} tag.
      *
      * @param string|Stringable $content Tag content.
+     * @param string|null $lang The document language.
      * @param array $attributes The tag attributes in terms of name-value pairs.
      */
     public static function html(string|Stringable $content = '', ?string $lang = null, array $attributes = []): Tag\Html